refactor: Payment 컴포넌트 디렉터리 구조를 개선한다
