+++ src/components/Payment/PaymentInfo/PaymentItem/PaymentItem.styled.jsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import { COLOR } from "../../../constants/style";
+import { COLOR } from "../../../../constants/style";
 
 export const PaymentItem = styled.div`
   display: flex;
